end of section 18

+++ src/app/messages/messages.component.ts
@@ -1,3 +1,4 @@
+import { ConfirmService } from './../_services/confirm.service';
 import { MessageService } from './../_services/message.service';
 import { MembersService } from 'src/app/_services/members.service';
 import { Pagination } from '../_models/pagination';
@@ -17,7 +18,7 @@ export class MessagesComponent implements OnInit {
   pageSize=5;
   loading=false;
 
-  constructor(private messageService : MessageService) { }
+  constructor(private messageService : MessageService, private confirmService : ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -33,9 +34,14 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id:number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+    this.confirmService.confirm('confirm delete message', 'this cannot be undone').subscribe(result => {
+      if(result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+        })
+      }
     })
+
   }
 
   pageChanged(event:any){
